[codedrop] billing and newAbonent added

+++ src/main/java/ru/learnup/nexigntask/callscdrplus/dto/charge/ChargeNumberBalanceDto.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class ChargeNumberBalanceDto {
 
     private final String phoneNumber;
-    private final int balance;
+    private final double balance;
 }
